+++ src/audiovideocontroller/DefaultAudioVideoController.ts
@@ -10,6 +10,7 @@ import AudioVideoObserver from '../audiovideoobserver/AudioVideoObserver';
 import DefaultBrowserBehavior from '../browserbehavior/DefaultBrowserBehavior';
 import ConnectionHealthData from '../connectionhealthpolicy/ConnectionHealthData';
 import SignalingAndMetricsConnectionMonitor from '../connectionmonitor/SignalingAndMetricsConnectionMonitor';
+import IdleMonitor from '../idlemonitor/DefaultIdleMonitor';
 import Logger from '../logger/Logger';
 import Maybe from '../maybe/Maybe';
 import MediaStreamBroker from '../mediastreambroker/MediaStreamBroker';
@@ -18,6 +19,7 @@ import MeetingSessionStatus from '../meetingsession/MeetingSessionStatus';
 import MeetingSessionStatusCode from '../meetingsession/MeetingSessionStatusCode';
 import MeetingSessionVideoAvailability from '../meetingsession/MeetingSessionVideoAvailability';
 import DefaultPingPong from '../pingpong/DefaultPingPong';
+import PingPong from '../pingpong/PingPong';
 import DefaultRealtimeController from '../realtimecontroller/DefaultRealtimeController';
 import RealtimeController from '../realtimecontroller/RealtimeController';
 import ReconnectController from '../reconnectcontroller/ReconnectController';
@@ -74,7 +76,7 @@ export default class DefaultAudioVideoController implements AudioVideoController
   private _mediaStreamBroker: MediaStreamBroker;
   private _reconnectController: ReconnectController;
   private _audioMixController: AudioMixController;
-
+  private pingPong: PingPong;
   private connectionHealthData = new ConnectionHealthData();
   private observerQueue: Set<AudioVideoObserver> = new Set<AudioVideoObserver>();
   private meetingSessionContext = new AudioVideoControllerState();
@@ -83,6 +85,7 @@ export default class DefaultAudioVideoController implements AudioVideoController
   private static MIN_VOLUME_DECIBELS = -42;
   private static MAX_VOLUME_DECIBELS = -14;
   private static PING_PONG_INTERVAL_MS = 10000;
+  private static MAX_IDLE_TIME_MS = 5 * 60 * 1000;
 
   constructor(
     configuration: MeetingSessionConfiguration,
@@ -216,16 +219,22 @@ export default class DefaultAudioVideoController implements AudioVideoController
     this.meetingSessionContext.videosToReceive = new DefaultVideoStreamIdSet();
     this.meetingSessionContext.videosPaused = new DefaultVideoStreamIdSet();
     this.meetingSessionContext.statsCollector = new DefaultStatsCollector(this, this.logger);
+    this.pingPong = new DefaultPingPong(
+      this.meetingSessionContext.signalingClient,
+      DefaultAudioVideoController.PING_PONG_INTERVAL_MS,
+      this.logger
+    );
+
+    this.meetingSessionContext.idleMonitor = new IdleMonitor(
+      this.pingPong,
+      DefaultAudioVideoController.MAX_IDLE_TIME_MS
+    );
     this.meetingSessionContext.connectionMonitor = new SignalingAndMetricsConnectionMonitor(
       this,
       this._realtimeController,
       this._videoTileController,
       this.connectionHealthData,
-      new DefaultPingPong(
-        this.meetingSessionContext.signalingClient,
-        DefaultAudioVideoController.PING_PONG_INTERVAL_MS,
-        this.logger
-      ),
+      this.pingPong,
       this.meetingSessionContext.statsCollector
     );
     this.meetingSessionContext.reconnectController = this._reconnectController;
@@ -241,9 +250,15 @@ export default class DefaultAudioVideoController implements AudioVideoController
 
     if (this._reconnectController.hasStartedConnectionAttempt()) {
       // This does not reset the reconnect deadline, but declare it's not the first connection.
-      this._reconnectController.startedConnectionAttempt(false);
+      this._reconnectController.startedConnectionAttempt(
+        false,
+        this.meetingSessionContext.idleMonitor
+      );
     } else {
-      this._reconnectController.startedConnectionAttempt(true);
+      this._reconnectController.startedConnectionAttempt(
+        true,
+        this.meetingSessionContext.idleMonitor
+      );
     }
 
     try {
@@ -495,7 +510,10 @@ export default class DefaultAudioVideoController implements AudioVideoController
 
   private async actionReconnect(): Promise<void> {
     if (!this._reconnectController.hasStartedConnectionAttempt()) {
-      this._reconnectController.startedConnectionAttempt(false);
+      this._reconnectController.startedConnectionAttempt(
+        false,
+        this.meetingSessionContext.idleMonitor
+      );
       this.forEachObserver(observer => {
         Maybe.of(observer.audioVideoDidStartConnecting).map(f => f.bind(observer)(true));
       });