+++ src/index.ts
@@ -47,6 +47,7 @@ import DefaultDOMWebSocketFactory from './domwebsocket/DefaultDOMWebSocketFactor
 import DefaultDeviceController from './devicecontroller/DefaultDeviceController';
 import DefaultDevicePixelRatioMonitor from './devicepixelratiomonitor/DefaultDevicePixelRatioMonitor';
 import DefaultDragObserver from './dragobserver/DefaultDragObserver';
+import DefaultIdleMonitor from './idlemonitor/DefaultIdleMonitor';
 import DefaultJPEGDecoderComponentFactory from './jpegdecoder/DefaultJPEGDecoderComponentFactory';
 import DefaultJPEGDecoderController from './jpegdecoder/controller/DefaultJPEGDecoderController';
 import DefaultJPEGDecoderInstance from './jpegdecoder/instance/DefaultJPEGDecoderInstance';
@@ -106,6 +107,7 @@ import FullJitterBackoff from './backoff/FullJitterBackoff';
 import FullJitterBackoffFactory from './backoff/FullJitterBackoffFactory';
 import FullJitterLimitedBackoff from './backoff/FullJitterLimitedBackoff';
 import GlobalMetricReport from './clientmetricreport/GlobalMetricReport';
+import IdleMonitor from './idlemonitor/IdleMonitor';
 import InitializeDefaultJPEGDecoderControllerTask from './task/InitializeDefaultJPEGDecoderControllerTask';
 import IntervalScheduler from './scheduler/IntervalScheduler';
 import JPEGDecoder from './jpegdecoder/webassembly/JPEGDecoder';
@@ -331,6 +333,7 @@ export {
   DefaultDeviceController,
   DefaultDevicePixelRatioMonitor,
   DefaultDragObserver,
+  DefaultIdleMonitor,
   DefaultJPEGDecoderComponentFactory,
   DefaultJPEGDecoderController,
   DefaultJPEGDecoderInstance,
@@ -390,6 +393,7 @@ export {
   FullJitterBackoffFactory,
   FullJitterLimitedBackoff,
   GlobalMetricReport,
+  IdleMonitor,
   InitializeDefaultJPEGDecoderControllerTask,
   IntervalScheduler,
   JPEGDecoder,