+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add basic component library setup
 - Add injectable session URL rewrite function to support proxies
 - Add POSTLogger for meeting sessions
+- Add sleep timeout condition to reconnect logic
 
 ### Changed
 - Stop content share if the media stream end
@@ -37,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix crash in demo app when click on screen share view
 - Fix integration test completion time writer
 - Fix the ping pong reconnection issue
+- Added sleep timeout condition to reconnect logic
 
 ## [1.1.0] - 2020-02-04